Fixing inconsistencies in the body model binding behavior in case of a missing property, or missing body. This handles 4 cases. 1. Body Null - Reference type FromBody property 2. Body Null - Value type FromBodyProperty 3. Reference type Property within the body null 4. Value type Property within the body null Now in cases 1-4 we use empty name as the prefix by default. What this means in all cases only the properties present in the body will be part of the model state dictionary key.
For example consider case 3
        public class Person
        {
            [FromBody]
            public Address Address { get; set; }
        }

        public class Address
        {
            [Required]
            public string Street { get; set; }

            public int Zip { get; set; }
        }
Request body { "Zip" : 12345 }
In this case the error key would be "Street" and not "prefix.Address.Street" (assuming there is a prefix because of additional metadata for the Person model).

The reason for omitting Address is since Address is not part of the body, a client cannot use ModelStateDictionary without knowing the model itself.
This is also consistent with the Json.Net behavior in case of value types being required and no value being provided(case4).